// src/types/api.ts
export interface components {
  schemas: {
// ...
    HealthResponseDto: {
      /**
       * @example ok
       * @enum {string}
       */
      status: 'ok' | 'degraded';
      /**
       * @example up
       * @enum {string}
       */
      database: 'up' | 'down';
      /**
       * Format: date-time
       * @example 2026-01-01T00:00:00.000Z
       */
      timestamp: string;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}